pyproject.toml
[project]
name = "self-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dacite>=1.9.2",
    "fastapi>=0.122.0",
    "hydra-core>=1.3.2",
    "langchain-community>=0.4.1",
    "langchain-qdrant>=1.1.0",
    "langchain[openai]>=1.2.0",
    "langgraph>=1.0.5",
    "loguru>=0.7.3",
    "pypdf>=6.4.2",
    "python-dotenv>=1.2.1",
    "qdrant-client>=1.16.2",
    "requests>=2.31.0",
    "rich>=14.2.0",
    "uvicorn>=0.38.0",
]
